+++ src/components/carbon-provider/carbon-provider.d.ts
@@ -1,0 +1,11 @@
+import * as React from "react";
+
+import { ThemeObject } from "../../style/themes/base";
+
+export interface CarbonProviderProps {
+  theme?: ThemeObject;
+  children: React.ReactNode;
+}
+
+declare function CarbonProvider(props: CarbonProviderProps): JSX.Element;
+export default CarbonProvider;