+++ src/components/pod/pod-manager.component.js
@@ -1,8 +1,18 @@
 import React, { useEffect, useState, useRef } from "react";
 import PropTypes from "prop-types";
 import PodContext from "./pod-context";
+import Logger from "../../utils/logger/logger";
+
+let deprecatedWarnTriggered = false;
 
 const PodManager = ({ children }) => {
+  if (!deprecatedWarnTriggered) {
+    deprecatedWarnTriggered = true;
+    // eslint-disable-next-line max-len
+    Logger.deprecate(
+      "`PodManager` component is deprecated and will soon be removed."
+    );
+  }
   const podManagerRef = useRef();
   const [heightOfTheLongestPod, setHeightOfTheLongestPod] = useState();