+++ src/components/select/filterable-select/filterable-select.component.js
@@ -391,7 +391,6 @@ const FilterableSelect = React.forwardRef(
       }
 
       isInputFocused.current = false;
-      setOpen(false);
 
       if (onBlur) {
         onBlur(event);