+++ src/components/select/multi-select/multi-select.component.js
@@ -300,7 +300,6 @@ const MultiSelect = React.forwardRef(
       }
 
       isInputFocused.current = false;
-      setOpenState(false);
 
       if (onBlur) {
         onBlur(event);